Extract rainbow, rainbowme into rainbow
